fm_etl_v3/requirements.txt
duckdb>=0.10.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
retrying>=1.3.4
